imu accelerometer drivers

+++ drivers/imu.h
@@ -3,9 +3,9 @@
 #include <stdint.h>
 void imu_init();
 
-int16_t get_x_accel();
-int16_t get_y_accel();
-int16_t get_z_accel();
+float get_x_accel();
+float get_y_accel();
+float get_z_accel();
 
 int16_t get_roll();
 int16_t get_yaw();
